requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.23.1
python-multipart==0.0.6
pillow==10.0.1
ultralytics==8.12.3
python-dotenv==1.0.0
aiofiles==23.1.0
numpy==1.26.2
